+++ tket/src/Predicates/PassGenerators.cpp
@@ -16,6 +16,7 @@
 
 #include <algorithm>
 #include <memory>
+#include <nlohmann/json_fwd.hpp>
 #include <sstream>
 #include <string>
 
@@ -480,14 +481,54 @@ PassPtr gen_cx_mapping_pass(
     bool delay_measures) {
   OpTypeSet gate_set = all_single_qubit_types();
   gate_set.insert(OpType::CX);
-  PassPtr rebase_pass =
-      gen_rebase_pass(gate_set, CircPool::CX(), CircPool::tk1_to_tk1);
+  PassPtr rebase_pass = gen_auto_rebase_pass(gate_set);
   PassPtr return_pass =
       rebase_pass >> gen_full_mapping_pass(arc, placement_ptr, config);
   if (delay_measures) return_pass = return_pass >> DelayMeasures();
   return_pass = return_pass >> rebase_pass >>
                 gen_decompose_routing_gates_to_cxs_pass(arc, directed_cx);
-  return return_pass;
+  Transform t{[=](Circuit& circ, std::shared_ptr<unit_bimaps_t> maps) {
+    // Relabel all two-qubit gates
+    CompilationUnit cu(circ);
+    bool changed = rebase_pass->apply(cu);
+    circ = cu.get_circ_ref();
+    // Now try placement, falling back on "LinePlacement" if "GraphPlacement"
+    // fails
+    try {
+      changed |= placement_ptr->place(circ, maps);
+    } catch (const std::runtime_error& e) {
+      std::stringstream ss;
+      ss << "PlacementPass failed with message: " << e.what()
+         << " Fall back to LinePlacement.";
+      tket_log()->warn(ss.str());
+      Placement::Ptr line_placement_ptr = std::make_shared<LinePlacement>(
+          placement_ptr->get_architecture_ref());
+      changed |= line_placement_ptr->place(circ, maps);
+    }
+    // Now route
+    MappingManager mm(std::make_shared<Architecture>(arc));
+    changed |= mm.route_circuit_with_maps(circ, config, maps);
+
+    // Now decompose routing gates to cx gates (this won't change maps)
+    CompilationUnit cu1(circ);
+    if (delay_measures) {
+      changed |= DelayMeasures()->apply(cu1);
+    }
+    changed |=
+        gen_decompose_routing_gates_to_cxs_pass(arc, directed_cx)->apply(cu1);
+    circ = cu1.get_circ_ref();
+    return changed;
+  }};
+  PassConditions conditions = return_pass->get_conditions();
+  nlohmann::json j;
+  j["name"] = "CXMappingPass";
+  j["architecture"] = arc;
+  j["placement"] = placement_ptr;
+  j["routing_config"] = config;
+  j["directed"] = directed_cx;
+  j["delay_measures"] = delay_measures;
+  return std::make_shared<StandardPass>(
+      conditions.first, t, conditions.second, j);
 }
 
 PassPtr gen_routing_pass(